+++ src/NoTimeForFishing/GlobalUsings.cs
@@ -1,5 +1,6 @@
 // Global using directives
 
+global using System;
 global using System.Reflection;
 global using BepInEx;
 global using BepInEx.Configuration;